fix a bug in compute_mask_indices() for wav2vec (#3266)
Summary:
# Before submitting

- [ ] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
- [x] Did you read the [contributor guideline](https://github.com/pytorch/fairseq/blob/master/CONTRIBUTING.md)?
- [ ] Did you make sure to update the docs?
- [x] Did you write any new necessary tests?

## What does this PR do?
Fixes a bug in the no_overlap case when computing mask indices for wav2vec

## PR review
Anyone in the community is free to review the PR once the tests have passed.
If we didn't discuss your PR in Github issues there's a high chance it will not be merged.

## Did you have fun?
Make sure you had fun coding �

Pull Request resolved: https://github.com/pytorch/fairseq/pull/3266

Reviewed By: arbabu123

Differential Revision: D35704063

Pulled By: alexeib

fbshipit-source-id: 3b77dc4cc50e539b57e6ad5f38f59eb975356adb